Add comments within comments feature and update tests
Commit Description:
This commit adds the ability to add comments within comments. The comments model has been modified to include a parent_comment_id attribute and a custom validation to ensure the replies are assigned with the post_id from the parent comment. The Comment model has been modified to accept nested attributes for replies. In addition, in this commit serializers were added to easily convert the objets to json and render the information in API format.
In addition to the new feature, the existing RSpec tests have been adjusted to reflect these changes. New tests have been added to test the creation and deletion of replies, as well as the validation of parent comment existence.
The README file has also been updated to include information about the new feature and how to use it.

Changes Made:
- Modified the comments model to include a parent_comment_id attribute and a custom validation to ensure that the replies have the post_id from parent_comment
- Modified the Comment model to accept nested attributes for replies.
- Adjusted the existing RSpec tests to reflect the changes and added new tests to test the creation and deletion of replies, as well as the validation of parent comment existence.
- Added serializers to render objects in json format
- Added annotation to models to display table schema information
- Added parent_comment_id to comments
- Implemented RSpec tests for serializers
- Updated the README file to include information about the new feature and how to use it.